requirements.txt
# HTTP Client
httpx>=0.27.0

# HTML Parsing
selectolax>=0.3.21

# Anthropic Claude API
anthropic>=0.21.0

# Environment Variables
python-dotenv>=1.0.0

# Database (optional: for async SQLite)
aiosqlite>=0.19.0
